% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_summary.R
\name{files_summary}
\alias{files_summary}
\title{files_summary}
\usage{
files_summary(root_in, export = FALSE, directory = NULL, plot_name = NULL)
}
\arguments{
\item{export}{}
}
\description{
files_summary
}
