% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plotdata.R
\name{import_plotdata}
\alias{import_plotdata}
\title{import_plotdata}
\usage{
import_plotdata(method, plot_name)
}
\arguments{
\item{method}{'plotsdatabase'}

\item{plot_name}{"mbalmayo001"}
}
\description{
import_plotdata
}
