% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XY_computation.R
\name{XY_computation}
\alias{XY_computation}
\title{XY_computation}
\usage{
XY_computation(
  Fieldplot_BDD_full,
  list_of_Lidar,
  plot_name,
  file_info,
  method = "jalon",
  directory
)
}
\arguments{
\item{method}{}
}
\description{
XY_computation
}
